vapp: refactor server
