+++ vapp/lib/asyncWrap.js
@@ -1,0 +1,5 @@
+module.exports = function asyncWrap (f) {
+  return function (res, req, next) {
+    f(res, req, next).catch(next);
+  };
+};
